feat: add 'Manage MCP servers' modal on chat creation
Replace the 'Add MCP server' button on the project's new-chat page with
'Manage MCP servers', opening a draft-mode WorkspaceMCPModal that lets
the user stage per-workspace enable/disable overrides and add new
servers before the workspace exists. Staged overrides are applied via
api.workspace.mcp.set immediately after workspace creation
(fire-and-forget so the first prompt is never blocked on remote
runtimes).

Refactors:
- Extract MCPAddServerForm and MCPOAuth helpers out of MCPSettingsSection
  so the new modal and Settings → MCP share one Add-server surface.
- Introduce shared workspaceMcpEffective helper (three-way precedence
  rule + toggle/hasAnyOverride) with 14 unit tests so the modal toggles
  and the overview count always agree.
- WorkspaceMCPModal becomes a discriminated-union dual-mode component
  ('workspace' = unchanged behavior, 'draft' = no live get/set, hides
  tool-allowlist UI, embeds MCPAddServerForm).
