+++ src/browser/features/Settings/Sections/MCPSettingsSection.tsx
@@ -1,13 +1,11 @@
-import React, { useCallback, useEffect, useRef, useState } from "react";
+import React, { useCallback, useEffect, useState } from "react";
 import { usePolicy } from "@/browser/contexts/PolicyContext";
 import { useAPI } from "@/browser/contexts/API";
 import {
   Trash2,
   Play,
   Loader2,
-  CheckCircle,
   XCircle,
-  Plus,
   Pencil,
   Check,
   X,
@@ -16,20 +14,12 @@ import {
   ChevronRight,
 } from "lucide-react";
 import { Button } from "@/browser/components/Button/Button";
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from "@/browser/components/SelectPrimitive/SelectPrimitive";
 import { createEditKeyHandler } from "@/browser/utils/ui/keybinds";
 import { Switch } from "@/browser/components/Switch/Switch";
 import { Tooltip, TooltipContent, TooltipTrigger } from "@/browser/components/Tooltip/Tooltip";
 import { cn } from "@/common/lib/utils";
 import { formatRelativeTime } from "@/browser/utils/ui/dateTime";
-import type { CachedMCPTestResult, MCPServerInfo, MCPServerTransport } from "@/common/types/mcp";
-import type { MCPOAuthPendingServerConfig } from "@/common/types/mcpOauth";
+import type { MCPServerInfo, MCPServerTransport } from "@/common/types/mcp";
 import { useMCPTestCache } from "@/browser/hooks/useMCPTestCache";
 import { MCPHeadersEditor } from "@/browser/components/MCPHeadersEditor/MCPHeadersEditor";
 import {
@@ -40,6 +30,14 @@ import {
 import { ToolSelector } from "@/browser/components/ToolSelector/ToolSelector";
 import { KebabMenu, type KebabMenuItem } from "@/browser/components/KebabMenu/KebabMenu";
 import { getErrorMessage } from "@/common/utils/errors";
+import { MCPAddServerForm } from "@/browser/components/MCPAddServerForm/MCPAddServerForm";
+import {
+  getMCPOAuthAPI,
+  getMCPOAuthLoginFlowMode,
+  MCPOAuthRequiredCallout,
+  useMCPOAuthLogin,
+  type MCPOAuthAuthStatus,
+} from "@/browser/components/MCPOAuth/MCPOAuth";
 
 /** Component for managing tool allowlist for a single MCP server */
 const ToolAllowlistSection: React.FC<{
@@ -168,356 +166,6 @@ const ToolAllowlistSection: React.FC<{
           />
         </div>
       )}
-    </div>
-  );
-};
-
-type MCPOAuthLoginStatus = "idle" | "starting" | "waiting" | "success" | "error";
-
-interface MCPOAuthAuthStatus {
-  serverUrl?: string;
-  isLoggedIn: boolean;
-  hasRefreshToken: boolean;
-  scope?: string;
-  updatedAtMs?: number;
-}
-
-type MCPOAuthAPI = NonNullable<ReturnType<typeof useAPI>["api"]>["mcpOauth"];
-
-function isRecord(value: unknown): value is Record<string, unknown> {
-  // In dev-server (browser) mode, the ORPC client can surface namespaces/procedures as Proxy
-  // functions (callable objects). Treat functions as record-like so runtime guards don't
-  // incorrectly report "OAuth is not available".
-  if (value === null) return false;
-  const type = typeof value;
-  return type === "object" || type === "function";
-}
-
-/**
- * Defensive runtime guard: `mcpOauth` may not exist when running against older backends
- * or in non-desktop environments. Treat OAuth as unavailable instead of surfacing raw exceptions.
- */
-function getMCPOAuthAPI(api: ReturnType<typeof useAPI>["api"]): MCPOAuthAPI | null {
-  if (!api) return null;
-
-  // Avoid direct property access since `api.mcpOauth` may be missing at runtime.
-  const maybeOauth: unknown = Reflect.get(api, "mcpOauth");
-  if (!isRecord(maybeOauth)) return null;
-
-  const requiredFns = ["getAuthStatus", "logout"] as const;
-
-  for (const fn of requiredFns) {
-    if (typeof maybeOauth[fn] !== "function") {
-      return null;
-    }
-  }
-
-  // Login flow support depends on whether the client can complete the callback.
-  const hasDesktopFlowFns =
-    typeof maybeOauth.startDesktopFlow === "function" &&
-    typeof maybeOauth.waitForDesktopFlow === "function" &&
-    typeof maybeOauth.cancelDesktopFlow === "function";
-
-  const hasServerFlowFns =
-    typeof maybeOauth.startServerFlow === "function" &&
-    typeof maybeOauth.waitForServerFlow === "function" &&
-    typeof maybeOauth.cancelServerFlow === "function";
-
-  if (!hasDesktopFlowFns && !hasServerFlowFns) {
-    return null;
-  }
-
-  return maybeOauth as unknown as MCPOAuthAPI;
-}
-
-type MCPOAuthLoginFlowMode = "desktop" | "server";
-
-function getMCPOAuthLoginFlowMode(input: {
-  isDesktop: boolean;
-  mcpOauthApi: MCPOAuthAPI | null;
-}): MCPOAuthLoginFlowMode | null {
-  const api = input.mcpOauthApi;
-  if (!api || !isRecord(api)) {
-    return null;
-  }
-
-  const hasDesktopFlowFns =
-    typeof api.startDesktopFlow === "function" &&
-    typeof api.waitForDesktopFlow === "function" &&
-    typeof api.cancelDesktopFlow === "function";
-
-  const hasServerFlowFns =
-    typeof api.startServerFlow === "function" &&
-    typeof api.waitForServerFlow === "function" &&
-    typeof api.cancelServerFlow === "function";
-
-  if (input.isDesktop) {
-    return hasDesktopFlowFns ? "desktop" : null;
-  }
-
-  return hasServerFlowFns ? "server" : null;
-}
-
-function useMCPOAuthLogin(input: {
-  api: ReturnType<typeof useAPI>["api"];
-  isDesktop: boolean;
-  serverName: string;
-  pendingServer?: MCPOAuthPendingServerConfig;
-  onSuccess?: () => void | Promise<void>;
-}) {
-  const { api, isDesktop, serverName, pendingServer, onSuccess } = input;
-  const loginAttemptRef = useRef(0);
-  const [flowId, setFlowId] = useState<string | null>(null);
-
-  const [loginStatus, setLoginStatus] = useState<MCPOAuthLoginStatus>("idle");
-  const [loginError, setLoginError] = useState<string | null>(null);
-
-  const loginInProgress = loginStatus === "starting" || loginStatus === "waiting";
-
-  const cancelLogin = useCallback(() => {
-    loginAttemptRef.current++;
-
-    const mcpOauthApi = getMCPOAuthAPI(api);
-    const loginFlowMode = getMCPOAuthLoginFlowMode({
-      isDesktop,
-      mcpOauthApi,
-    });
-
-    if (mcpOauthApi && flowId && loginFlowMode === "desktop") {
-      void mcpOauthApi.cancelDesktopFlow({ flowId });
-    }
-
-    if (mcpOauthApi && flowId && loginFlowMode === "server") {
-      void mcpOauthApi.cancelServerFlow({ flowId });
-    }
-
-    setFlowId(null);
-    setLoginStatus("idle");
-    setLoginError(null);
-  }, [api, flowId, isDesktop]);
-
-  const startLogin = useCallback(async () => {
-    const attempt = ++loginAttemptRef.current;
-
-    try {
-      setLoginError(null);
-      setFlowId(null);
-
-      if (!api) {
-        setLoginStatus("error");
-        setLoginError("Mux API not connected.");
-        return;
-      }
-
-      if (!serverName.trim()) {
-        setLoginStatus("error");
-        setLoginError("Server name is required to start OAuth login.");
-        return;
-      }
-
-      const mcpOauthApi = getMCPOAuthAPI(api);
-      if (!mcpOauthApi) {
-        setLoginStatus("error");
-        setLoginError("OAuth is not available in this environment.");
-        return;
-      }
-
-      const loginFlowMode = getMCPOAuthLoginFlowMode({
-        isDesktop,
-        mcpOauthApi,
-      });
-      if (!loginFlowMode) {
-        setLoginStatus("error");
-        setLoginError("OAuth login is not available in this environment.");
-        return;
-      }
-
-      setLoginStatus("starting");
-
-      const startResult =
-        loginFlowMode === "desktop"
-          ? await mcpOauthApi.startDesktopFlow({ serverName, pendingServer })
-          : await mcpOauthApi.startServerFlow({ serverName, pendingServer });
-
-      if (attempt !== loginAttemptRef.current) {
-        if (startResult.success) {
-          if (loginFlowMode === "desktop") {
-            void mcpOauthApi.cancelDesktopFlow({ flowId: startResult.data.flowId });
-          } else {
-            void mcpOauthApi.cancelServerFlow({ flowId: startResult.data.flowId });
-          }
-        }
-        return;
-      }
-
-      if (!startResult.success) {
-        setLoginStatus("error");
-        setLoginError(startResult.error);
-        return;
-      }
-
-      const { flowId: nextFlowId, authorizeUrl } = startResult.data;
-      setFlowId(nextFlowId);
-      setLoginStatus("waiting");
-
-      // Desktop main process intercepts external window.open() calls and routes them via shell.openExternal.
-      // In browser mode, this opens a new tab/window.
-      //
-      // NOTE: In some browsers (especially when using `noopener`), `window.open()` may return null even when
-      // the tab opens successfully. Do not treat a null return value as a failure signal; keep the OAuth flow
-      // alive and show guidance to the user while we wait.
-      try {
-        window.open(authorizeUrl, "_blank", "noopener");
-      } catch {
-        // Popups can be blocked or restricted by the browser. The user can cancel and retry after allowing
-        // popups; we intentionally do not auto-cancel the server flow here.
-      }
-
-      if (attempt !== loginAttemptRef.current) {
-        return;
-      }
-
-      const waitResult =
-        loginFlowMode === "desktop"
-          ? await mcpOauthApi.waitForDesktopFlow({ flowId: nextFlowId })
-          : await mcpOauthApi.waitForServerFlow({ flowId: nextFlowId });
-
-      if (attempt !== loginAttemptRef.current) {
-        return;
-      }
-
-      if (waitResult.success) {
-        setLoginStatus("success");
-        await onSuccess?.();
-        return;
-      }
-
-      setLoginStatus("error");
-      setLoginError(waitResult.error);
-    } catch (err) {
-      if (attempt !== loginAttemptRef.current) {
-        return;
-      }
-
-      const message = getErrorMessage(err);
-      setLoginStatus("error");
-      setLoginError(message);
-    }
-  }, [api, isDesktop, onSuccess, pendingServer, serverName]);
-
-  return {
-    loginStatus,
-    loginError,
-    loginInProgress,
-    startLogin,
-    cancelLogin,
-  };
-}
-
-const MCPOAuthRequiredCallout: React.FC<{
-  serverName: string;
-  pendingServer?: MCPOAuthPendingServerConfig;
-  disabledReason?: string;
-  onLoginSuccess?: () => void | Promise<void>;
-}> = ({ serverName, pendingServer, disabledReason, onLoginSuccess }) => {
-  const { api } = useAPI();
-  const isDesktop = !!window.api;
-
-  const { loginStatus, loginError, loginInProgress, startLogin, cancelLogin } = useMCPOAuthLogin({
-    api,
-    isDesktop,
-    serverName,
-    pendingServer,
-    onSuccess: onLoginSuccess,
-  });
-
-  const mcpOauthApi = getMCPOAuthAPI(api);
-  const loginFlowMode = getMCPOAuthLoginFlowMode({
-    isDesktop,
-    mcpOauthApi,
-  });
-
-  const disabledTitle =
-    disabledReason ??
-    (!api
-      ? "Mux API not connected"
-      : !mcpOauthApi
-        ? "OAuth is not available in this environment."
-        : !loginFlowMode
-          ? isDesktop
-            ? "OAuth login is not available in this environment."
-            : "OAuth login is only available in the desktop app."
-          : undefined);
-
-  const loginDisabled = Boolean(disabledReason) || !api || !loginFlowMode || loginInProgress;
-
-  const loginButton = (
-    <Button
-      size="sm"
-      onClick={() => {
-        void startLogin();
-      }}
-      disabled={loginDisabled}
-      aria-label="Login via OAuth"
-    >
-      {loginInProgress ? (
-        <>
-          <Loader2 className="h-3.5 w-3.5 animate-spin" />
-          Waiting for login...
-        </>
-      ) : (
-        "Login via OAuth"
-      )}
-    </Button>
-  );
-
-  return (
-    <div className="bg-warning/10 border-warning/30 text-warning rounded-md border px-3 py-2 text-xs">
-      <div className="flex items-start justify-between gap-3">
-        <div className="min-w-0">
-          <p className="font-medium">This server requires OAuth.</p>
-          {disabledReason && <p className="text-muted mt-0.5">{disabledReason}</p>}
-
-          {loginStatus === "waiting" && (
-            <>
-              <p className="text-muted mt-0.5">
-                Finish the login flow in your browser, then return here.
-              </p>
-              {!isDesktop && (
-                <p className="text-muted mt-0.5">
-                  If a new tab didn&apos;t open, your browser may have blocked the popup. Allow
-                  popups and try again.
-                </p>
-              )}
-            </>
-          )}
-
-          {loginStatus === "success" && <p className="text-muted mt-0.5">Logged in.</p>}
-
-          {loginStatus === "error" && loginError && (
-            <p className="text-destructive mt-0.5">OAuth error: {loginError}</p>
-          )}
-        </div>
-
-        <div className="flex shrink-0 items-center gap-2">
-          {disabledTitle ? (
-            <Tooltip>
-              <TooltipTrigger asChild>
-                <span className="inline-flex">{loginButton}</span>
-              </TooltipTrigger>
-              <TooltipContent side="top">{disabledTitle}</TooltipContent>
-            </Tooltip>
-          ) : (
-            loginButton
-          )}
-
-          {loginStatus === "waiting" && (
-            <Button variant="secondary" size="sm" onClick={cancelLogin}>
-              Cancel
-            </Button>
-          )}
-        </div>
-      </div>
     </div>
   );
 };
@@ -746,50 +394,6 @@ export const MCPSettingsSection: React.FC = () => {
     headersRows: MCPHeaderRow[];
   }
 
-  // Add form state
-
-  // Ensure the "Add server" transport select always points to a policy-allowed value.
-  useEffect(() => {
-    if (!mcpAllowUserDefined) {
-      return;
-    }
-
-    const isAllowed = (transport: MCPServerTransport): boolean => {
-      if (transport === "stdio") {
-        return mcpAllowUserDefined.stdio;
-      }
-
-      return mcpAllowUserDefined.remote;
-    };
-
-    setNewServer((prev) => {
-      if (isAllowed(prev.transport)) {
-        return prev;
-      }
-
-      const fallback: MCPServerTransport | null = mcpAllowUserDefined.stdio
-        ? "stdio"
-        : mcpAllowUserDefined.remote
-          ? "http"
-          : null;
-
-      if (!fallback) {
-        return prev;
-      }
-
-      return { ...prev, transport: fallback, value: "", headersRows: [] };
-    });
-  }, [mcpAllowUserDefined]);
-  const [newServer, setNewServer] = useState<EditableServer>({
-    name: "",
-    transport: "stdio",
-    value: "",
-    headersRows: [],
-  });
-  const [addingServer, setAddingServer] = useState(false);
-  const [testingNew, setTestingNew] = useState(false);
-  const [newTestResult, setNewTestResult] = useState<CachedMCPTestResult | null>(null);
-
   // Edit state
   const [editing, setEditing] = useState<EditableServer | null>(null);
   const [savingEdit, setSavingEdit] = useState(false);
@@ -836,11 +440,6 @@ export const MCPSettingsSection: React.FC = () => {
   useEffect(() => {
     void refresh();
   }, [refresh]);
-
-  // Clear new-server test result when transport/value/headers change
-  useEffect(() => {
-    setNewTestResult(null);
-  }, [newServer.transport, newServer.value, newServer.headersRows]);
 
   const handleRemove = useCallback(
     async (name: string) => {
@@ -918,129 +517,6 @@ export const MCPSettingsSection: React.FC = () => {
   const serverDisplayValue = (entry: MCPServerInfo): string =>
     entry.transport === "stdio" ? entry.command : entry.url;
 
-  const handleTestNewServer = useCallback(async () => {
-    if (!api || !newServer.value.trim()) return;
-    setTestingNew(true);
-    setNewTestResult(null);
-
-    try {
-      const { headers, validation } =
-        newServer.transport === "stdio"
-          ? { headers: undefined, validation: { errors: [], warnings: [] } }
-          : mcpHeaderRowsToRecord(newServer.headersRows, {
-              knownSecretKeys: new Set(globalSecretKeys),
-            });
-
-      if (validation.errors.length > 0) {
-        throw new Error(validation.errors[0]);
-      }
-
-      const pendingName = newServer.name.trim();
-
-      const result = await api.mcp.test({
-        ...(newServer.transport === "stdio"
-          ? { command: newServer.value.trim() }
-          : {
-              ...(pendingName ? { name: pendingName } : {}),
-              transport: newServer.transport,
-              url: newServer.value.trim(),
-              headers,
-            }),
-      });
-
-      setNewTestResult({ result, testedAt: Date.now() });
-    } catch (err) {
-      setNewTestResult({
-        result: { success: false, error: err instanceof Error ? err.message : "Test failed" },
-        testedAt: Date.now(),
-      });
-    } finally {
-      setTestingNew(false);
-    }
-  }, [
-    api,
-    newServer.name,
-    newServer.transport,
-    newServer.value,
-    newServer.headersRows,
-    globalSecretKeys,
-  ]);
-
-  const handleAddServer = useCallback(async () => {
-    if (!api || !newServer.name.trim() || !newServer.value.trim()) return;
-
-    const serverName = newServer.name.trim();
-    const serverTransport = newServer.transport;
-    const serverValue = newServer.value.trim();
-    const serverHeadersRows = newServer.headersRows;
-    const existingTestResult = newTestResult;
-
-    setAddingServer(true);
-    setError(null);
-
-    try {
-      const { headers, validation } =
-        serverTransport === "stdio"
-          ? { headers: undefined, validation: { errors: [], warnings: [] } }
-          : mcpHeaderRowsToRecord(serverHeadersRows, {
-              knownSecretKeys: new Set(globalSecretKeys),
-            });
-
-      if (validation.errors.length > 0) {
-        throw new Error(validation.errors[0]);
-      }
-
-      const result = await api.mcp.add({
-        name: serverName,
-        ...(serverTransport === "stdio"
-          ? { transport: "stdio", command: serverValue }
-          : {
-              transport: serverTransport,
-              url: serverValue,
-              headers,
-            }),
-      });
-
-      if (!result.success) {
-        setError(result.error ?? "Failed to add MCP server");
-        return;
-      }
-
-      setNewServer({ name: "", transport: "stdio", value: "", headersRows: [] });
-      setNewTestResult(null);
-      await refresh();
-
-      // For stdio, avoid running arbitrary user-provided commands automatically.
-      if (serverTransport === "stdio") {
-        if (existingTestResult?.result.success) {
-          cacheTestResult(serverName, existingTestResult.result);
-        }
-        return;
-      }
-
-      // For remote servers, always run a test immediately after adding so OAuth-required servers can
-      // surface an OAuth callout without requiring a manual Test click.
-      setTestingServer(serverName);
-      try {
-        const testResult = await api.mcp.test({
-          name: serverName,
-        });
-        cacheTestResult(serverName, testResult);
-      } catch (err) {
-        cacheTestResult(serverName, {
-          success: false,
-          error: err instanceof Error ? err.message : "Test failed",
-        });
-      } finally {
-        setTestingServer(null);
-      }
-    } catch (err) {
-      setError(err instanceof Error ? err.message : "Failed to add MCP server");
-    } finally {
-      setAddingServer(false);
-    }
-  }, [api, newServer, newTestResult, refresh, cacheTestResult, globalSecretKeys]);
-
   const handleStartEdit = useCallback((name: string, entry: MCPServerInfo) => {
     setEditing({
       name,
@@ -1096,22 +572,6 @@ export const MCPSettingsSection: React.FC = () => {
       setSavingEdit(false);
     }
   }, [api, editing, refresh, clearTestResult, globalSecretKeys]);
-
-  const newHeadersValidation =
-    newServer.transport === "stdio"
-      ? { errors: [], warnings: [] }
-      : mcpHeaderRowsToRecord(newServer.headersRows, {
-          knownSecretKeys: new Set(globalSecretKeys),
-        }).validation;
-
-  const canAdd =
-    newServer.name.trim().length > 0 &&
-    newServer.value.trim().length > 0 &&
-    (newServer.transport === "stdio" || newHeadersValidation.errors.length === 0);
-
-  const canTest =
-    newServer.value.trim().length > 0 &&
-    (newServer.transport === "stdio" || newHeadersValidation.errors.length === 0);
 
   const editHeadersValidation =
     editing && editing.transport !== "stdio"
@@ -1405,206 +865,7 @@ export const MCPSettingsSection: React.FC = () => {
             </div>
 
             {/* Add server form */}
-            <details className="group mt-3">
-              <summary className="text-accent hover:text-accent/80 flex cursor-pointer list-none items-center gap-1 text-sm font-medium">
-                <ChevronRight className="h-4 w-4 transition-transform group-open:rotate-90" />
-                Add server
-              </summary>
-              <div className="border-border-medium bg-background-secondary mt-2 space-y-3 rounded-md border p-3">
-                <div>
-                  <label htmlFor="server-name" className="text-muted mb-1 block text-xs">
-                    Name
-                  </label>
-                  <input
-                    id="server-name"
-                    type="text"
-                    placeholder="e.g., memory"
-                    value={newServer.name}
-                    onChange={(e) => setNewServer((prev) => ({ ...prev, name: e.target.value }))}
-                    className="bg-modal-bg border-border-medium focus:border-accent w-full rounded border px-2 py-1.5 text-sm focus:outline-none"
-                  />
-                </div>
-
-                <div>
-                  <label className="text-muted mb-1 block text-xs">Transport</label>
-                  <Select
-                    value={newServer.transport}
-                    onValueChange={(value) =>
-                      setNewServer((prev) => ({
-                        ...prev,
-                        transport: value as MCPServerTransport,
-                        value: "",
-                        headersRows: [],
-                      }))
-                    }
-                  >
-                    <SelectTrigger className="border-border-medium bg-modal-bg h-8 w-full text-sm">
-                      <SelectValue />
-                    </SelectTrigger>
-                    <SelectContent>
-                      {mcpAllowUserDefined?.stdio !== false && (
-                        <SelectItem value="stdio">Stdio</SelectItem>
-                      )}
-                      {mcpAllowUserDefined?.remote !== false && (
-                        <>
-                          <SelectItem value="http">HTTP (Streamable)</SelectItem>
-                          <SelectItem value="sse">SSE (Legacy)</SelectItem>
-                          <SelectItem value="auto">Auto (HTTP → SSE)</SelectItem>
-                        </>
-                      )}
-                    </SelectContent>
-                  </Select>
-                </div>
-
-                <div>
-                  <label htmlFor="server-value" className="text-muted mb-1 block text-xs">
-                    {newServer.transport === "stdio" ? "Command" : "URL"}
-                  </label>
-                  <input
-                    id="server-value"
-                    type="text"
-                    placeholder={
-                      newServer.transport === "stdio"
-                        ? "e.g., npx -y @modelcontextprotocol/server-memory"
-                        : "e.g., http://localhost:3333/mcp"
-                    }
-                    value={newServer.value}
-                    onChange={(e) => setNewServer((prev) => ({ ...prev, value: e.target.value }))}
-                    spellCheck={false}
-                    className="bg-modal-bg border-border-medium focus:border-accent w-full rounded border px-2 py-1.5 font-mono text-sm focus:outline-none"
-                  />
-                </div>
-
-                {newServer.transport !== "stdio" && (
-                  <div>
-                    <label className="text-muted mb-1 block text-xs">HTTP headers (optional)</label>
-                    <MCPHeadersEditor
-                      rows={newServer.headersRows}
-                      onChange={(rows) =>
-                        setNewServer((prev) => ({
-                          ...prev,
-                          headersRows: rows,
-                        }))
-                      }
-                      secretKeys={globalSecretKeys}
-                      disabled={addingServer || testingNew}
-                    />
-                  </div>
-                )}
-
-                {/* Test result */}
-                {newTestResult && (
-                  <div
-                    className={cn(
-                      "flex items-start gap-2 rounded-md px-3 py-2 text-sm",
-                      newTestResult.result.success
-                        ? "bg-green-500/10 text-green-500"
-                        : "bg-destructive/10 text-destructive"
-                    )}
-                  >
-                    {newTestResult.result.success ? (
-                      <>
-                        <CheckCircle className="mt-0.5 h-4 w-4 shrink-0" />
-                        <div>
-                          <span className="font-medium">
-                            Connected — {newTestResult.result.tools.length} tools
-                          </span>
-                          {newTestResult.result.tools.length > 0 && (
-                            <p className="mt-0.5 text-xs opacity-80">
-                              {newTestResult.result.tools.join(", ")}
-                            </p>
-                          )}
-                        </div>
-                      </>
-                    ) : (
-                      <>
-                        <XCircle className="mt-0.5 h-4 w-4 shrink-0" />
-                        <span>{newTestResult.result.error}</span>
-                      </>
-                    )}
-                  </div>
-                )}
-
-                {newTestResult &&
-                  !newTestResult.result.success &&
-                  newTestResult.result.oauthChallenge && (
-                    <div className="mt-2">
-                      <MCPOAuthRequiredCallout
-                        serverName={newServer.name.trim()}
-                        pendingServer={(() => {
-                          const pendingName = newServer.name.trim();
-                          if (!pendingName) {
-                            return undefined;
-                          }
-
-                          // If the server already exists in config, prefer that config for OAuth.
-                          const existing = servers[pendingName];
-                          if (existing) {
-                            return undefined;
-                          }
-
-                          if (newServer.transport === "stdio") {
-                            return undefined;
-                          }
-
-                          const url = newServer.value.trim();
-                          if (!url) {
-                            return undefined;
-                          }
-
-                          return { transport: newServer.transport, url };
-                        })()}
-                        disabledReason={(() => {
-                          const pendingName = newServer.name.trim();
-                          if (!pendingName) {
-                            return "Enter a server name to enable OAuth login.";
-                          }
-
-                          const existing = servers[pendingName];
-
-                          const transport = existing?.transport ?? newServer.transport;
-                          if (transport === "stdio") {
-                            return "OAuth login is only supported for remote (http/sse) MCP servers.";
-                          }
-
-                          return undefined;
-                        })()}
-                        onLoginSuccess={async () => {
-                          setMcpOauthRefreshNonce((prev) => prev + 1);
-                          await handleTestNewServer();
-                        }}
-                      />
-                    </div>
-                  )}
-                <div className="flex gap-2 pt-1">
-                  <Button
-                    variant="outline"
-                    size="sm"
-                    onClick={() => void handleTestNewServer()}
-                    disabled={!canTest || testingNew}
-                  >
-                    {testingNew ? (
-                      <Loader2 className="h-3.5 w-3.5 animate-spin" />
-                    ) : (
-                      <Play className="h-3.5 w-3.5" />
-                    )}
-                    {testingNew ? "Testing…" : "Test"}
-                  </Button>
-                  <Button
-                    size="sm"
-                    onClick={() => void handleAddServer()}
-                    disabled={!canAdd || addingServer}
-                  >
-                    {addingServer ? (
-                      <Loader2 className="h-3.5 w-3.5 animate-spin" />
-                    ) : (
-                      <Plus className="h-3.5 w-3.5" />
-                    )}
-                    {addingServer ? "Adding…" : "Add"}
-                  </Button>
-                </div>
-              </div>
-            </details>
+            <MCPAddServerForm existingServers={servers} onAdded={() => refresh()} />
           </>
         )}
       </div>
